components: Move Carousel to components
Generalizes the Carousel component and move it to the components package
to use it on the discovery page.

https://phabricator.endlessm.com/T31937
