+++ packages/template-ui/src/constants.js
@@ -1,2 +1,1 @@
 export const SEARCH_MAX_RESULTS = 64;
-export const CAROUSEL_INTERVAL = 10000;
