+++ packages/template-ui/src/views/Home.vue
@@ -11,7 +11,7 @@
     </div>
 
     <div v-else>
-      <Carousel />
+      <Carousel :nodes="carouselNodes" />
       <b-container>
         <hr>
       </b-container>
@@ -51,11 +51,20 @@
 
 <script>
 import { mapState, mapGetters } from 'vuex';
+import _ from 'underscore';
 
 export default {
   name: 'Home',
   computed: {
-    ...mapState(['section', 'loading', 'cardColumns', 'mediaQuality']),
+    ...mapState([
+      'nodes',
+      'carouselNodeIds',
+      'carouselSlideNumber',
+      'section',
+      'loading',
+      'cardColumns',
+      'mediaQuality',
+    ]),
     ...mapGetters({
       mainSections: 'mainSections',
       getAssetURL: 'getAssetURL',
@@ -69,6 +78,25 @@ export default {
         return null;
       }
       return this.section.children.filter((n) => n.kind !== 'topic') || null;
+    },
+    carouselNodes() {
+      if (this.carouselNodeIds.length) {
+        return this.carouselNodesFixed(this.carouselNodeIds);
+      }
+
+      return this.carouselNodesRandom(this.carouselSlideNumber);
+    },
+  },
+  methods: {
+    carouselNodesRandom(n) {
+      // Get n random nodes that are not topic:
+      const possibleNodes = this.nodes.filter((node) => node.kind !== 'topic');
+      return _.sample(possibleNodes, n);
+    },
+    carouselNodesFixed(nodeIds) {
+      return nodeIds.map((n) => (
+        this.nodes.find((m) => m.id === n.id)
+      ));
     },
   },
 };
